backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.1.0
pymupdf==1.26.1
anthropic==0.50.0
httpx==0.28.1
google-cloud-storage==2.19.0
google-api-python-client==2.166.0
psycopg2-binary==2.9.9
PyJWT>=2.8.0
slowapi>=0.1.9
python-json-logger>=2.0.7
